+++ ElgrosLib/DataModels/Category.cs
@@ -6,7 +6,7 @@
 
         public string Name { get { return _name; } }
 
-        internal Category(long id, string name) : base(id)
+        internal Category(int id, string name) : base(id)
         {
             _name = name;
         }
