+++ ElgrosLib/DataModels/SubCategory.cs
@@ -8,7 +8,7 @@
         public string Name { get { return _name; } }
         public int CategoryId { get { return _categoryId; } }
 
-        public SubCategory(long id, string name, int cagegoryId) : base(id)
+        public SubCategory(int id, string name, int cagegoryId) : base(id)
         {
             _name = name;
             _categoryId = cagegoryId;
